Make CreateWorldItem() in WorldItemManager take float values for finer pos offsets

+++ Assets/Scripts/HUD/InventorySystem/WorldItemManager.cs
@@ -63,7 +63,7 @@ public class WorldItemManager : MonoBehaviour
     /*
      * Creates a world item
      */
-    public void CreateWorldItem(int world_item_id, int item_id, int amt, int pos_x, int pos_y)
+    public void CreateWorldItem(int world_item_id, int item_id, int amt, float pos_x, float pos_y)
     {
         Item item = _item_manager.FetchItemById(item_id);
         GameObject _item = Instantiate(world_item);
